pyproject.toml
[project]
name = "adaptive-self-consistency"
version = "0.1.0"
description = "A-SC2: Adaptive Self-Consistency with Confidence- and Cost-Regularized Early Exit"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "wandb>=0.16.0",
    "datasets>=2.16.0",
    "google-generativeai>=0.3.0",
    "tiktoken>=0.5.2",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "matplotlib>=3.8.0",
    "pyyaml>=6.0",
    "omegaconf>=2.3.0",
    "tqdm>=4.66.0",
]